nci: improved buffer handling
Each buffer is now a separate Frame object that is stuffed into the body of
the main view, rather than a separate UI.  This provides much better cleanup
when cycling through buffers, and is just cleaner in general.

Each buffer Frame now carries it's own header info as well.  The status bar
is still handled by the main UI.